+++ src/__internal__/sticky-footer/index.js
@@ -1,0 +1,1 @@
+export { default } from "./sticky-footer.component";