+++ src/components/drawer/drawer.spec.js
@@ -3,19 +3,22 @@
 import React from "react";
 import { act } from "react-dom/test-utils";
 import { mount as enzymeMount, shallow } from "enzyme";
+
 import Drawer from "./drawer.component";
-import { assertStyleMatch } from "../../../__spec_helper__/test-utils";
-import guid from "../../../utils/helpers/guid/guid";
+import { assertStyleMatch } from "../../__spec_helper__/test-utils";
+import guid from "../../utils/helpers/guid/guid";
 import {
   StyledDrawerSidebar,
   StyledDrawerContent,
   StyledDrawerChildren,
   StyledSidebarTitle,
   StyledButton,
+  StyledSidebarHeader,
 } from "./drawer.style";
-import { noThemeSnapshot } from "../../../__spec_helper__/enzyme-snapshot-helper";
+import { noThemeSnapshot } from "../../__spec_helper__/enzyme-snapshot-helper";
+import StickyFooter from "../../__internal__/sticky-footer";
 
-jest.mock("../../../utils/helpers/guid");
+jest.mock("../../utils/helpers/guid");
 guid.mockImplementation(() => "guid-123");
 
 let container = null;
@@ -46,19 +49,17 @@ const render = (props, renderer = mount) => {
 };
 
 const getElements = (wrapper) => {
-  const cw = wrapper;
-
-  if (!cw) {
+  if (!wrapper) {
     return {};
   }
 
   return {
-    drawer: cw.find(Drawer),
-    sidebar: cw.find(StyledDrawerSidebar),
-    content: cw.find(StyledDrawerContent),
-    children: cw.find(StyledDrawerChildren),
-    button: cw.find(StyledButton),
-    title: cw.find(StyledSidebarTitle),
+    drawer: wrapper.find(Drawer),
+    sidebar: wrapper.find(StyledDrawerSidebar),
+    content: wrapper.find(StyledDrawerContent),
+    children: wrapper.find(StyledDrawerChildren),
+    button: wrapper.find(StyledButton),
+    title: wrapper.find(StyledSidebarTitle),
   };
 };
 
@@ -117,7 +118,9 @@ describe("Drawer", () => {
       const { sidebar } = getElements(wrapper);
       assertStyleMatch(
         {
-          overflow: "auto",
+          display: "flex",
+          flexDirection: "column",
+          flex: "1 1 0%",
         },
         sidebar
       );
@@ -232,6 +235,83 @@ describe("Drawer", () => {
       wrapper.update();
       const { content } = getElements(wrapper);
       expect(content.childAt(0).hasClass("closed")).toBeTruthy();
+    });
+
+    describe("with the stickyHeader prop set", () => {
+      describe("when expanded", () => {
+        it("should add the correct styles", () => {
+          const wrapper = render({
+            stickyHeader: true,
+            showControls: true,
+            title: "Test title",
+          });
+
+          assertStyleMatch(
+            {
+              position: "sticky",
+              top: "0",
+              borderBottom: "1px solid #ccd6db",
+            },
+            wrapper.find(StyledSidebarHeader)
+          );
+
+          assertStyleMatch(
+            {
+              position: "sticky",
+              top: "0",
+              borderBottom: "1px solid #ccd6db",
+            },
+            wrapper.find(StyledSidebarHeader)
+          );
+        });
+      });
+
+      describe("when closed", () => {
+        it("should add the correct styles", () => {
+          const wrapper = render({
+            stickyHeader: true,
+            showControls: true,
+            title: "Test title",
+            expanded: false,
+          });
+
+          assertStyleMatch(
+            {
+              position: "sticky",
+              top: "0",
+              borderBottom: undefined,
+            },
+            wrapper.find(StyledSidebarHeader)
+          );
+        });
+      });
+    });
+
+    describe("with the footer prop set", () => {
+      describe("when stickyFooter prop is false", () => {
+        it("should not be sticky", () => {
+          const wrapper = render({
+            footer: <div>Some footer content</div>,
+          });
+
+          expect(wrapper.find(StickyFooter).props().disableSticky).toEqual(
+            true
+          );
+        });
+      });
+
+      describe("when stickyFooter prop is true", () => {
+        it("should be sticky", () => {
+          const wrapper = render({
+            footer: <div>Some footer content</div>,
+            stickyFooter: true,
+          });
+
+          expect(wrapper.find(StickyFooter).props().disableSticky).toEqual(
+            false
+          );
+        });
+      });
     });
 
     describe("invariant", () => {